requirements.txt
# Core dependencies
tensorflow==2.15.0
streamlit==1.32.0
numpy==1.24.3
pillow==10.1.0

# Streamlit extensions
streamlit-extras==0.3.5

# Visualization (for development/analysis)
matplotlib==3.8.2
seaborn==0.13.0

# Data processing
pandas==2.1.3
scikit-learn==1.3.2